Split up encryption modules further
